bitbake/toaster-requirements.txt
Django>1.8,<1.9
argparse==1.2.1
wsgiref==0.1.2
beautifulsoup4>=4.4.0
djangorestframework>=3.3.3
django-filter>=0.13.0
requests>=2.10.0
